Disable snapshots by default; 64-bit-safe indexing
Add a scaling study guide and improve scaling/robustness:

- Add SCALING_STUDY_COMMANDS.md with usage/examples for the Makefile scaling sweep.
- Makefile: set SCALING_SNAP_EVERY default to 0 to avoid excessive I/O, introduce SCALING_BLOCK_X/Y defaults, include block_x/block_y in results/scaling.csv, pass block dims to cuda/hybrid runs, and abort if build steps fail.
- heat.h: make IDX macro 64-bit-safe by casting operands to size_t to avoid overflow on large grids.
- main.c: accept snap-every == 0, only write snapshots when snap_every > 0, use size_t for buffer sizes/loops and for allocation arithmetic, and declare cudaDeviceProp with struct qualifier.
- stencil_cuda.cu: convert linear indexing to size_t-safe arithmetic across kernels and tile loads/stores.

These changes reduce I/O for large scaling runs and fix indexing/size arithmetic to be safe on very large grids; also surface CUDA block shape in CSV results for easier analysis.

+++ src/heat.h
@@ -71,8 +71,8 @@ typedef enum { MAT_FR4 = 0, MAT_COPPER = 1, MAT_TIM = 2, MAT_SILICON = 3 } mater
     } while (0)
 #endif
 
-/* 2D row-major indexing helper: linear_index = j * nx + i. */
-#define IDX(g, i, j) ((j) * (g)->nx + (i))
+/* 2D row-major indexing helper with 64-bit-safe arithmetic. */
+#define IDX(g, i, j) ((size_t)(j) * (size_t)(g)->nx + (size_t)(i))
 
 /* Shared simulation state used by CPU, CUDA, and MPI paths. */
 typedef struct {
